moved constants file to app folder, adjusted paths to point to that folder instead of previous
